update flow login facebook for website

+++ application/controllers/frontend/Customer.php
@@ -258,7 +258,7 @@ class Customer extends MY_Controller
             $customer = $this->Mcustomers->checkExist(0, $postData['customer_email']);
             $customerId = 0;
             if (count($customer) > 0) $customerId = $customer['id'];
-            $message = 'Successfully register account';
+            $message = $this->lang->line('mobile_successfully_register_account');
             $customer['language_id'] = 1;
 
             if ($customerId == 0) {
@@ -273,11 +273,30 @@ class Customer extends MY_Controller
             }
             $flag = $this->Mcustomers->save($postData, $customerId);
             if ($flag > 0) {
-                $customer = $this->Mcustomers->get($flag);
-                $customer['language_name'] = $customer['language_id'] == 0 ? 'english' : $this->Mconstants->languageCodes[$customer['language_id']];
-                $this->load->helper('cookie');
-                $this->input->set_cookie($this->configValueCookie('customer', json_encode($customer), '3600'));
-                echo json_encode(array('code' => 1, 'message' => $message, 'data' => $flag));
+                $checkExist = true;
+
+                if($postData['login_type_id'] == 1) {
+                    $customerNew = $this->Mcustomers->get($flag);
+                    if(empty($customerNew['customer_email'])) {
+                        $this->session->set_userdata('customer_update_email', $flag);
+                        $checkExist = false;
+                        echo json_encode(array('code' => 3, 'data' => $flag, 'url' => base_url('customer/update-email?customer='.$flag)));die;
+                    }else{
+                        $checkExist = true;
+                        $ext = explode('@', $postData['customer_email']);
+                        if(count($ext) > 1 && $ext[1] == 'facebook.com') {
+                            $this->session->set_userdata('customer_update_email', $flag);
+                            echo json_encode(array('code' => 3, 'data' => $flag, 'url' => base_url('customer/update-email?customer='.$flag)));die;
+                        }
+                    }
+                }
+                if($checkExist == true){
+                    $customer = $this->Mcustomers->get($flag);
+                    $customer['language_name'] = $customer['language_id'] == 0 ? 'english' : $this->Mconstants->languageCodes[$customer['language_id']];
+                    $this->load->helper('cookie');
+                    $this->input->set_cookie($this->configValueCookie('customer', json_encode($customer), '3600'));
+                    echo json_encode(array('code' => 1, 'message' => $message, 'data' => $flag));
+                }
             } else echo json_encode(array('code' => 0, 'message' => ERROR_COMMON_MESSAGE));
         } catch (\Throwable $th) {
             echo json_encode(array('code' => -2, 'message' => ERROR_COMMON_MESSAGE));
@@ -972,6 +991,74 @@ class Customer extends MY_Controller
             $this->session->set_flashdata('notice_message', ERROR_COMMON_MESSAGE);
             $this->session->set_flashdata('notice_type', 'error');
             redirect(base_url('login.html'));
+        }
+    }
+    
+    public function updateEmail(){
+        $this->loadModel(array('Mconfigs', 'Mcustomers'));
+
+        /**
+         * Commons data
+         */
+        $data = $this->commonDataCustomer($this->lang->line('update_customer_email'));
+        $data['activeMenu'] = "";
+        /**
+         * Commons data
+         */
+
+        $this->load->view('frontend/customer/update-email', $data);
+    }
+
+    public function submitUpdateEMail(){
+        try {
+            $postData = $this->arrayFromPost(array('customer_email'));
+
+            if (!empty($postData['customer_email'])) {
+                $this->load->model('Mcustomers');
+
+                $chkExistEmail = $this->Mcustomers->getFieldValue(array('customer_email' => $postData['customer_email'], 'customer_status_id >' => 0), 'id', 0);
+                if($chkExistEmail){
+                    echo json_encode(array('code' => 0, 'message' => $this->lang->line('221121_email_already_exist')));die;
+                }
+
+                $customerId = $this->rsession->get('customer_update_email');
+                if ($customerId > 0) {
+                    $dataUpdate = array(
+                        'customer_email' => $postData['customer_email']
+                    );
+                    $customerId = $this->Mcustomers->save($dataUpdate, $customerId);
+                    if($customerId > 0){
+                        $this->rsession->delete('customer_update_email');
+
+                        $customerInfo = $this->Mcustomers->get($customerId);
+
+                        $token_active = guidV4('update-email');
+
+                        $flag = $this->Mcustomers->save(array('token_reset' => $token_active), $customerId);
+
+                        $this->load->model('Memailqueue');
+                        $name = !empty($customerInfo['customer_first_name']) ? $postData['customer_first_name'] : $postData['customer_email'];
+                        $dataEmail = array(
+                            'name' => $name,
+                            'email_to' => $customerInfo['customer_email'],
+                            'email_to_name' => $customerInfo['customer_email'],
+                            'token' => $token_active
+                        );
+                        $this->Memailqueue->createEmail($dataEmail, 99);
+
+                        echo json_encode(array('code' => 1, 'message' => $this->lang->line('update_email_success')));die;
+                    }else{
+                        echo json_encode(array('code' => 0, 'message' => $this->lang->line('update_email_failed')));die;
+                    }
+                } else {
+                    echo json_encode(array('code' => 0, 'message' => $this->lang->line('customer_not_exist')));die;
+                }
+                
+            } else {
+                echo json_encode(array('code' => 0, 'message' => $this->lang->line('221121_please_enter_your_email')));die;
+            }
+        } catch (Exception $e) {
+            echo json_encode(array('code' => 0, 'message' => ERROR_COMMON_MESSAGE));die;
         }
     }
 }
